Random background colors and new properties.
Added two new functions: getRandomColor() and randomizeBackgroundColor(). Together they build an RGB color code and use it as the background color for the body of the document every time a quote is printed.

I also added two new optional properties to the quote objects: learnMore and link. They will be used to add a link to certain quotes that will allow the user to watch a trailer or buy a book that inspired the quote.